[KMCompiler][TLERaw] Move the hook registration to tle raw cuda runtime
